Client (#5)
* login done! (#1)

This pull request completes the frontend development of the Login page for the Inventory Management System (IMS). The login interface is now fully designed, responsive, and integrated with client-side validation to improve user experience and usability.

* Dashboard 1.0 (#4)

This pull request introduces the Dashboard User Interface for the Inventory Management System (IMS). The dashboard provides a centralized overview of inventory-related information with a clean, modern, and responsive design to enhance usability and navigation.